+++ python/GafferSceneUI/TweakPlugValueWidget.py
@@ -42,6 +42,8 @@ import Gaffer
 import GafferUI
 import GafferScene
 
+from Qt import QtWidgets
+
 # Widget for TweakPlug, which is used to build tweak nodes such as ShaderTweaks
 # and CameraTweaks.  Shows a value plug that you can use to specify a tweak value, along with
 # a target parameter name, an enabled plug, and a mode.  The mode can be "Replace",
@@ -70,7 +72,11 @@ class TweakPlugValueWidget( GafferUI.PlugValueWidget ) :
 			verticalAlignment = GafferUI.Label.VerticalAlignment.Top,
 		)
 
-		self.__row.append( GafferUI.PlugValueWidget.create( plug["mode"] ) )
+		modeWidget = GafferUI.PlugValueWidget.create( plug["mode"] )
+		modeWidget._qtWidget().setFixedWidth( 80 )
+		modeWidget._qtWidget().layout().setSizeConstraint( QtWidgets.QLayout.SetDefaultConstraint )
+		self.__row.append( modeWidget )
+
 		self.__row.append( GafferUI.PlugValueWidget.create( plug["value"] ), expand = True )
 
 		self._updateFromPlug()
@@ -105,6 +111,14 @@ class TweakPlugValueWidget( GafferUI.PlugValueWidget ) :
 
 		for w in self.__row :
 			w.setReadOnly( readOnly )
+
+	def setNameVisible( self, visible ) :
+
+		self.__row[0].setVisible( visible )
+
+	def getNameVisible( self ) :
+
+		return self.__row[0].getVisible()
 
 	def _updateFromPlug( self ) :
 
@@ -189,6 +203,18 @@ def __noduleLabel( plug ) :
 
 	return name or plug.getName()
 
+Gaffer.Metadata.registerValue( GafferScene.TweakPlug, "nodule:type", "GafferUI::CompoundNodule" )
+Gaffer.Metadata.registerValue( GafferScene.TweakPlug, "*", "nodule:type", "" )
+Gaffer.Metadata.registerValue( GafferScene.TweakPlug, "value", "nodule:type", "GafferUI::StandardNodule" )
+Gaffer.Metadata.registerValue( GafferScene.TweakPlug, "noduleLayout:label", __noduleLabel )
+Gaffer.Metadata.registerValue( GafferScene.TweakPlug, "value", "noduleLayout:label", __noduleLabel )
+
+# Spreadsheet Interoperability
+# ============================
+
+Gaffer.Metadata.registerValue( GafferScene.TweakPlug, "spreadsheet:plugMenu:includeAsAncestor", True )
+Gaffer.Metadata.registerValue( GafferScene.TweakPlug, "spreadsheet:plugMenu:ancestorLabel", "Tweak" )
+
 def __spreadsheetColumnName( plug ) :
 
 	tweakPlug = plug.parent()
@@ -207,10 +233,28 @@ def __spreadsheetColumnName( plug ) :
 	else :
 		return plug.getName()
 
-Gaffer.Metadata.registerValue( GafferScene.TweakPlug, "nodule:type", "GafferUI::CompoundNodule" )
-Gaffer.Metadata.registerValue( GafferScene.TweakPlug, "*", "nodule:type", "" )
-Gaffer.Metadata.registerValue( GafferScene.TweakPlug, "value", "nodule:type", "GafferUI::StandardNodule" )
-Gaffer.Metadata.registerValue( GafferScene.TweakPlug, "noduleLayout:label", __noduleLabel )
-Gaffer.Metadata.registerValue( GafferScene.TweakPlug, "value", "noduleLayout:label", __noduleLabel )
 Gaffer.Metadata.registerValue( GafferScene.TweakPlug, "*", "spreadsheet:columnName", __spreadsheetColumnName )
 
+def __spreadsheetFormatter( plug, forToolTip ) :
+
+	result = ""
+	if "enabled" in plug.parent() :
+		result = "On, " if plug["enabled"].getValue() else "Off, "
+
+	result += str( GafferScene.TweakPlug.Mode.values[plug["mode"].getValue()] )
+
+	value = GafferUI.SpreadsheetUI.formatValue( plug["value"], forToolTip )
+	separator = " : \n" if forToolTip and "\n" in value else " : "
+	result += separator + value
+
+	return result
+
+GafferUI.SpreadsheetUI.registerValueFormatter( GafferScene.TweakPlug, __spreadsheetFormatter )
+
+def __spreadsheetValueWidget( plug ) :
+
+	w = TweakPlugValueWidget( plug )
+	w.setNameVisible( False )
+	return w
+
+GafferUI.SpreadsheetUI.registerValueWidget( GafferScene.TweakPlug, __spreadsheetValueWidget )